Re #204: Hide stack widgets we aren't showing
Part of the slow-down is because GTK invalidates and recalculates ALL
of the widgets in the stack. But if we hide them then it doesn't need
to do that. So when the transition ends, hide all widgets we're not
showing.

Currently seems to behave the same, but with faster focus switching.