Re #204: Add and remove pages when navigating
This is the harshest version of removing pages when we're not using
them. We only keep extra pages in the stack until we're done with them.
But the pages are still created and keep state, so they don't lose
positions etc

Also now pass the main window around so that we don't assume a
base/parent when it might not be set.

This won't work as-is with the libhandy redesign (because we need to
keep all of the main tabs as children so we populate the switcher).